Move VoyageCheckYourAnswers page into correct folder
