+++ src/AppRouter.jsx
@@ -81,7 +81,7 @@ const ContactUs = lazy(() => import('./pages/Regulatory/ContactUs'));
 // Downloadable templates pages
 const Templates = lazy(() => import('./pages/NavPages/Templates'));
 // Voyage pages
-const VoyageCheckYourAnswers = lazy(() => import('./pages/Voyage/VoyageCheckYourAnswers'));
+const VoyageCheckYourAnswers = lazy(() => import('./pages/Voyage/VoyageCheckYourAnswers/VoyageCheckYourAnswers'));
 const VoyageCrew = lazy(() => import('./pages/Voyage/VoyageCrew'));
 const VoyageDeleteDraftCheck = lazy(() => import('./pages/Voyage/VoyageDeleteDraftCheck'));
 const VoyageGeneralDeclaration = lazy(() => import('./pages/Voyage/VoyageGeneralDeclaration'));
